requirements.txt
apipkg==1.4
cffi==1.5.0
coverage==4.0.3
docopt==0.6.2
execnet==1.4.1
flake8==2.5.1
freetype-py==1.0.2
glfw-cffi==0.1.8
mccabe==0.3.1
numpy==1.10.4
pep8==1.7.0
py==1.4.31
pycparser==2.14
pyflakes==1.0.0
PyOpenGL==3.1.0
pytest==2.8.7
pytest-cache==1.0
pytest-cov==2.2.0
pytest-flake8==0.1
pytest-xdist==1.13.1
wheel==0.26.0
